Carve launch, match_config, probe out of game_loop
Stage 1 of slice 9 (host-restructure cleanup): create top-level public
modules `pyrat_host::{launch, match_config, probe}` so callers stop
importing through `game_loop`. The internals are unchanged; this is a
file move plus one targeted codec swap.

- `launch.rs`: lifted from `game_loop/launch.rs`; absorbs `BotConfig`
  from `game_loop/config.rs` (it's a launch concern, not a generic
  config).
- `match_config.rs`: new home for `build_match_config` and its tests.
- `probe.rs`: lifted from `game_loop/probe.rs`; swaps the legacy
  `crate::session::extract_bot_packet` for the canonical
  `pyrat_protocol::extract_bot_msg`. Probing is host orchestration
  (used by the GUI's bot-config panel), not a `game_loop` concern.

`game_loop/mod.rs` becomes a thin re-export shim — legacy callers
(`tools/bot-check`, integration tests) keep building unchanged. The
shim is dropped wholesale in slice 9 stage 5 once headless / GUI /
bot-check have moved to the new import paths.

No behavior change. 138 host tests + headless e2e green. fmt + clippy
clean across pyrat-host, pyrat-headless, pyrat-check.
